Calm Before Storm

+++ quizapp/quizapps/views.py
@@ -28,7 +28,7 @@ def quiz(request, quiz_id):
     return render(request, 'quizapps/quiz.html', context)
 
 
-@login_required
+@ login_required
 def question(request, question_id):
     question = Question.objects.get(id=question_id)
 
@@ -44,12 +44,12 @@ def question(request, question_id):
     return render(request, 'quizapps/question.html', context)
 
 
-@login_required
+@ login_required
 def results(request, question_id):  # Maybe add quiz id of results replace to quiz id
     global score  # Must make it reset for each quiz
     choice_id = request.POST['Choice']
     choice = Choice.objects.get(id=choice_id)
-    #totalquestions = len(list(Question.objects.get(id=question_id)))
+    # totalquestions = len(list(Question.objects.get(id=question_id)))
     totalquestions = len(list(Question.objects.get(id=question_id).quiz.question_set.all()))
     if choice.correct_choice:
         score += 1
@@ -64,7 +64,7 @@ def results(request, question_id):  # Maybe add quiz id of results replace to qu
 # create a view for making questions
 
 
-@login_required
+@ login_required
 def resultsquiz(request, quiz_id):
     global score
     quiz = Quizs.objects.get(id=quiz_id)
@@ -73,101 +73,113 @@ def resultsquiz(request, quiz_id):
     return render(request, 'quizapps/resultsquiz.html', context)
 
 
-@login_required
+@ login_required
 def create_quiz(request):
-    if request.method != 'POST':
-        form = QuizForm()
-    else:
-        form = QuizForm(data=request.POST)
-        if form.is_valid():
-            form.save()
-        return redirect('quizapps:index')
-    context = {'form': form}
-    return render(request, 'quizapps/create_quiz.html', context)
-
-
-@login_required
-def create_question(request, quiz_id):  # Maybe make it such that it is for a quiz
-    quiz = Quizs.objects.get(id=quiz_id)
-    if quiz.owner != request.user:
-        raise Http404
-    ChoiceFormSet = modelformset_factory(Choice, fields=('choice_text', 'correct_choice'), extra=4)
-    if request.method != 'POST':
-        # qz_form = QuizForm()
-        q_form = QuestionForm()
-        c_form = ChoiceFormSet(queryset=Choice.objects.none())
-        # for edit use queryset such that it filters to the question__id = questoin.id
-    else:
-        # qz_form = QuizForm(data=request.POST)
-        q_form = QuestionForm(data=request.POST)
-        c_form = ChoiceFormSet(data=request.POST)
-
-        # if all([qz_form.is_valid(), q_form.is_valid(), c_form.is_valid()]):
-        if all([q_form.is_valid(), c_form.is_valid()]):
-            # qz = qz_form.save()
-            q = q_form.save(commit=False)
-            c = c_form.save(commit=False)
-            # Assigning Quiz to Question
-            q.quiz = Quizs.objects.get(id=quiz_id)
-            print(c)
-            # non commit false returns None
-            q.save()
-            for form in c:
-                # print(instance.question, q)
-                # print(form)
-                form.question = q
-                form.save()
-            return redirect('quizapps:quiz', quiz_id)
-
-            # c_form.save()
-    quiz = Quizs.objects.get(id=quiz_id)
-    # context = {'qz_form': qz_form, 'q_form': q_form, 'c_form': c_form}
-    context = {'q_form': q_form, 'c_form': c_form, 'quiz': quiz}
-    return render(request, 'quizapps/create_question.html', context)
-
-
-@login_required
-def edit_question(request, question_id):  # Maybe make it such that it is for a quiz
-    question = Question.objects.get(id=question_id)
-    if question.quiz.owner != request.user:
-        raise Http404
-    ChoiceFormSet = modelformset_factory(Choice, fields=(
-        'choice_text', 'correct_choice'), extra=4, max_num=4)
-    if request.method != 'POST':
-        # qz_form = QuizForm()
-        q_form = QuestionForm(instance=question)
-        c_form = ChoiceFormSet(queryset=Choice.objects.filter(question=question))
-        # for edit use queryset such that it filters to the question__id = questoin.id
-    else:
-        # qz_form = QuizForm(data=request.POST)
-        q_form = QuestionForm(data=request.POST, instance=question)
-        c_form = ChoiceFormSet(data=request.POST, queryset=Choice.objects.filter(question=question))
-        # print(c_form)
-        # if all([qz_form.is_valid(), q_form.is_valid(), c_form.is_valid()]):
-        if all([q_form.is_valid(), c_form.is_valid()]):
-            # qz = qz_form.save()
-            q = q_form.save(commit=False)
-            c = c_form.save(commit=False)
-            # q.quiz = Quizs.objects.get(id=quiz_id)
-            # print(c)
-            q.save()
-            for form in c:
-                # print(instance.question, q)
-                # print(form)
-                form.question = q
-                form.save()
-            return redirect('quizapps:quizview', question.quiz.id)
+    if request.user.userprofile.role == "TR":
+        if request.method != 'POST':
+            form = QuizForm()
         else:
-            print(c_form.errors)
+            form = QuizForm(data=request.POST)
+            if form.is_valid():
+                form.save()
+            return redirect('quizapps:index')
+        context = {'form': form}
+        return render(request, 'quizapps/create_quiz.html', context)
+    else:
+        return HttpResponse("No Student Can Create Quiz!")
 
-            # c_form.save()
-    # quiz = Quizs.objects.get(id=quiz_id)
-    # context = {'qz_form': qz_form, 'q_form': q_form, 'c_form': c_form}
-    context = {'q_form': q_form, 'c_form': c_form, 'question': question}
-    return render(request, 'quizapps/edit_question.html', context)
+
+@ login_required
+def create_question(request, quiz_id):  # Maybe make it such that it is for a quiz
+    if request.user.userprofile.role == "TR":
+        quiz = Quizs.objects.get(id=quiz_id)
+        if quiz.owner != request.user:
+            raise Http404
+        ChoiceFormSet = modelformset_factory(
+            Choice, fields=('choice_text', 'correct_choice'), extra=4)
+        if request.method != 'POST':
+            # qz_form = QuizForm()
+            q_form = QuestionForm()
+            c_form = ChoiceFormSet(queryset=Choice.objects.none())
+            # for edit use queryset such that it filters to the question__id = questoin.id
+        else:
+            # qz_form = QuizForm(data=request.POST)
+            q_form = QuestionForm(data=request.POST)
+            c_form = ChoiceFormSet(data=request.POST)
+
+            # if all([qz_form.is_valid(), q_form.is_valid(), c_form.is_valid()]):
+            if all([q_form.is_valid(), c_form.is_valid()]):
+                # qz = qz_form.save()
+                q = q_form.save(commit=False)
+                c = c_form.save(commit=False)
+                # Assigning Quiz to Question
+                q.quiz = Quizs.objects.get(id=quiz_id)
+                print(c)
+                # non commit false returns None
+                q.save()  # save the non commit instance not the form
+                for form in c:
+                    # print(instance.question, q)
+                    # print(form)
+                    form.question = q
+                    form.save()
+                return redirect('quizapps:quiz', quiz_id)
+
+                # c_form.save()
+        quiz = Quizs.objects.get(id=quiz_id)
+        # context = {'qz_form': qz_form, 'q_form': q_form, 'c_form': c_form}
+        context = {'q_form': q_form, 'c_form': c_form, 'quiz': quiz}
+        return render(request, 'quizapps/create_question.html', context)
+
+    else:
+        return HttpResponse("No Student Can Create Question!")
 
 
-@login_required
+@ login_required
+def edit_question(request, question_id):  # Maybe make it such that it is for a quiz
+    if request.user.userprofile.role == "TR":
+        question = Question.objects.get(id=question_id)
+        if question.quiz.owner != request.user:
+            raise Http404
+        ChoiceFormSet = modelformset_factory(Choice, fields=(
+            'choice_text', 'correct_choice'), extra=4, max_num=4)
+        if request.method != 'POST':
+            # qz_form = QuizForm()
+            q_form = QuestionForm(instance=question)
+            c_form = ChoiceFormSet(queryset=Choice.objects.filter(question=question))
+            # for edit use queryset such that it filters to the question__id = questoin.id
+        else:
+            # qz_form = QuizForm(data=request.POST)
+            q_form = QuestionForm(data=request.POST, instance=question)
+            c_form = ChoiceFormSet(
+                data=request.POST, queryset=Choice.objects.filter(question=question))
+            # print(c_form)
+            # if all([qz_form.is_valid(), q_form.is_valid(), c_form.is_valid()]):
+            if all([q_form.is_valid(), c_form.is_valid()]):
+                # qz = qz_form.save()
+                q = q_form.save(commit=False)
+                c = c_form.save(commit=False)
+                # q.quiz = Quizs.objects.get(id=quiz_id)
+                # print(c)
+                q.save()
+                for form in c:
+                    # print(instance.question, q)
+                    # print(form)
+                    form.question = q
+                    form.save()
+                return redirect('quizapps:quizview', question.quiz.id)
+            else:
+                print(c_form.errors)
+
+                # c_form.save()
+        # quiz = Quizs.objects.get(id=quiz_id)
+        # context = {'qz_form': qz_form, 'q_form': q_form, 'c_form': c_form}
+        context = {'q_form': q_form, 'c_form': c_form, 'question': question}
+        return render(request, 'quizapps/edit_question.html', context)
+    else:
+        return HttpResponse("No Student Can Edit Question!")
+
+
+@ login_required
 def quizview(request, quiz_id):
     quiz = Quizs.objects.get(id=quiz_id)
     if quiz.owner != request.user:
@@ -177,15 +189,65 @@ def quizview(request, quiz_id):
     return render(request, 'quizapps/quizview.html', context)
 
 
-@login_required
+@ login_required
 def delete_question(request, question_id):
-    question = Question.objects.get(id=question_id)
-    question.delete()
-    return redirect('quizapps:quizview', question.quiz.id)
+    if request.user.userprofile.role == "TR":
+        question = Question.objects.get(id=question_id)
+        if question.quiz.owner != request.user:
+            raise Http404
+        question.delete()
+        return redirect('quizapps:quizview', question.quiz.id)
+    else:
+        return HttpResponse("No Student Can Delete Question!")
 
 
-@login_required
+@ login_required
 def delete_quiz(request, quiz_id):
-    quiz = Quizs.objects.get(id=quiz_id)
-    quiz.delete()
-    return redirect('quizapps:index')
+    if request.user.userprofile.role == "TR":
+        quiz = Quizs.objects.get(id=quiz_id)
+        if quiz.owner != request.user:
+            raise Http404
+        quiz.delete()
+        return redirect('quizapps:index')
+    else:
+        return HttpResponse("No Student Can Delete Quiz!")
+
+
+@ login_required
+def assign_quiz(request, quiz_id):
+    if request.user.userprofile.role == "TR":
+        quiz = Quizs.objects.get(id=quiz_id)
+
+        # Genrating Users
+        CHOICES = []
+        for user in User.objects.all():
+            if user.userprofile.role == "ST":
+                CHOICES.append((user.id, user))
+
+        if quiz.owner != request.user:
+            raise Http404
+
+        if request.method != 'POST':
+            form = MultipleUserForm()
+            form.fields['MultipleUser'].choices = CHOICES
+        else:
+            form = MultipleUserForm()
+            form.fields['MultipleUser'].choices = CHOICES
+
+            for id in request.POST['MultipleUser']:
+                quiz.owner = User.objects.get(id=id)
+
+            # print(request.POST)
+
+            #form = MultipleUserForm(data=request.POST)
+            # if form.is_valid():
+            #form = GeeksForm(data=request.POST)
+            # form.save()
+            # print(request.POST)
+            # return redirect('quizapps:quiz', quiz_id)
+
+        context = {'form': form, 'quiz': quiz}
+        return render(request, 'quizapps/assign_quiz.html', context)
+    else:
+
+        return HttpResponse("No Student Can Assign Quiz!")
